requirements.txt
wget
pushbullet.py==0.9.1
